[template] Contract refactor handling function data
